Match ktrexfloorswitch_update charge level masking
Type the charge level local as u8 (not u32) so the post-increment
re-masks to a byte at each use (#20/#58), matching target's
addi/clrlwi burst. 97.57 -> 97.65.

+++ src/main/dll/DR/dll_0251_ktrexfloorswitch.c
@@ -169,7 +169,7 @@ void ktrexfloorswitch_update(int obj)
     ObjTextureRuntimeSlot* tex;
     int* player;
     int moved;
-    u32 level;
+    u8 level;
     int scroll;
     f32 vecA[3];
     f32 vecB[3];
